fix bugs

+++ module/pan_run_hite_single.py
@@ -115,6 +115,7 @@ if __name__ == "__main__":
     work_dir = os.path.abspath(work_dir)
     shared_prev_TE = args.shared_prev_TE
 
+    reference = os.path.abspath(reference)
     # 处理输出目录
     output_dir = os.path.abspath(args.output_dir)
     os.makedirs(output_dir, exist_ok=True)
@@ -137,11 +138,11 @@ if __name__ == "__main__":
     except Exception as e:
         # 如果出现异常，打印错误信息并删除临时目录
         print(f"An error occurred: {e}")
-        # if os.path.exists(temp_dir):
-        #     shutil.rmtree(temp_dir)
+        if os.path.exists(temp_dir) and debug != 1:
+            shutil.rmtree(temp_dir)
         raise  # 重新抛出异常，以便上层代码可以处理
 
     else:
         # 如果没有异常，删除临时目录
-        if os.path.exists(temp_dir):
+        if os.path.exists(temp_dir) and debug != 1:
             shutil.rmtree(temp_dir)
